refactor: Introduce CommandOutputFormatter and refactor command handling for consistent output
- Added `CommandOutputFormatter` to standardize command output across text and JSON formats, reducing duplication in command handlers.
- Refactored command handling in `CommandRegistry` to utilize the new formatter for success, error, and validation messages.
- Created `ConfigCommandBuilder` and `RecordCommandBuilder` to improve maintainability and separation of concerns in command construction.
