+++ src/Infrastructure/Cli/CommandRegistry.cs
@@ -12,6 +12,7 @@ using TenSecondTom.Features.Shell.Services;
 using TenSecondTom.Features.ThisWeek.Handlers;
 using TenSecondTom.Features.Today.Handlers;
 using TenSecondTom.Infrastructure.Auth;
+using TenSecondTom.Infrastructure.Configuration;
 using TenSecondTom.Shared.Constants;
 using TenSecondTom.Shared.OutputFormatters;
 using AuthLoginHandler = TenSecondTom.Features.Auth.Handlers.LoginCommandHandler;
@@ -26,12 +27,6 @@ namespace TenSecondTom.Infrastructure.Cli;
 public static class CommandRegistry
 {
     private static readonly string[] QuitAliases = ["exit"];
-
-    private static readonly System.Text.Json.JsonSerializerOptions SnakeCaseJsonOptions = new()
-    {
-        WriteIndented = true,
-        PropertyNamingPolicy = System.Text.Json.JsonNamingPolicy.SnakeCaseLower
-    };
 
     /// <summary>
     /// Builds and configures the root command with all subcommands.
@@ -437,203 +432,7 @@ public static class CommandRegistry
 
     private static Command BuildRecordCommand(IServiceProvider serviceProvider, Option<bool> jsonOutputOption)
     {
-        var recordCommand = new Command("record", "Record audio with transcription and save to recording/ directory");
-
-        // Add options
-        var sttOption = new Option<string?>("--stt")
-        {
-            Description = "STT engine selection: auto (default), local, or openai."
-        };
-
-        recordCommand.Options.Add(sttOption);
-        recordCommand.Options.Add(jsonOutputOption);
-
-        // Set action
-        recordCommand.SetAction(async (parseResult) =>
-        {
-            bool jsonOutput = parseResult.GetValue(jsonOutputOption);
-            string? stt = parseResult.GetValue(sttOption);
-
-            // Get AudioConfiguration to read default SttProvider and timeout
-            var audioConfig = serviceProvider.GetService<IOptions<Configuration.AudioConfiguration>>()?.Value
-                ?? new Configuration.AudioConfiguration();
-            
-            // Parse STT selection - use configured default or fall back to Auto
-            var sttSelection = TenSecondTom.Features.Audio.Models.SttSelection.Auto;
-            
-            // If no --stt flag provided, use configuration default
-            if (string.IsNullOrWhiteSpace(stt))
-            {
-                if (audioConfig.SttProvider is not null)
-                {
-                    if (!Enum.TryParse<TenSecondTom.Features.Audio.Models.SttSelection>(audioConfig.SttProvider, ignoreCase: true, out sttSelection))
-                    {
-                        // Invalid config value, fall back to Auto
-                        sttSelection = TenSecondTom.Features.Audio.Models.SttSelection.Auto;
-                    }
-                }
-            }
-            else
-            {
-                // --stt flag provided, parse and validate it
-                if (!Enum.TryParse<TenSecondTom.Features.Audio.Models.SttSelection>(stt, ignoreCase: true, out sttSelection))
-                {
-                    if (jsonOutput)
-                    {
-                        Console.WriteLine(System.Text.Json.JsonSerializer.Serialize(new
-                        {
-                            success = false,
-                            error = $"Invalid STT selection: {stt}. Valid options: auto, local, openai"
-                        }));
-                    }
-                    else
-                    {
-                        AnsiConsole.MarkupLine($"[red]Invalid STT selection:[/] {stt.EscapeMarkup()}");
-                        AnsiConsole.MarkupLine("[dim]Valid options: auto, local, openai[/]");
-                    }
-                    return 1;
-                }
-            }
-
-            // Authenticate first (record command creates data that will be used by authenticated commands)
-            var authService = serviceProvider.GetRequiredService<IAuthenticationService>();
-
-            // Show warning if using mock authentication (only in non-JSON mode)
-            if (!jsonOutput && authService is MockAuthenticationService)
-            {
-                AnsiConsole.MarkupLine("[yellow]⚠ Development Mode: Authentication bypassed[/]");
-                AnsiConsole.WriteLine();
-            }
-
-            var authResult = await AuthenticationHelper.EnsureAuthenticatedAsync(
-                authService,
-                CommandNames.Record,
-                jsonOutput,
-                CancellationToken.None);
-
-            if (!authResult.IsSuccess)
-            {
-                return 1; // Authentication failed
-            }
-
-            // Validate audio configuration
-            var audioValidator = serviceProvider.GetRequiredService<TenSecondTom.Features.Audio.Services.IAudioConfigurationValidator>();
-            var audioConfigResult = TenSecondTom.Features.Audio.AudioConfigurationHelper.EnsureAudioConfigured(
-                audioValidator,
-                audioConfig,
-                CommandNames.Record,
-                jsonOutput);
-
-            if (!audioConfigResult.IsSuccess)
-            {
-                return 1; // Audio configuration incomplete
-            }
-
-            // Get handler
-            var handler = serviceProvider.GetService<TenSecondTom.Features.Audio.Handlers.RecordCommandHandler>();
-            if (handler is null)
-            {
-                if (jsonOutput)
-                {
-                    Console.WriteLine(System.Text.Json.JsonSerializer.Serialize(new
-                    {
-                        success = false,
-                        error = "Record functionality is unavailable - handler not registered in DI container."
-                    }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine("[red]Record unavailable:[/] handler not registered. Ensure AddFeatureAudioServices() was called.");
-                }
-                return 1;
-            }
-
-            // Show recording prompt (unless in JSON mode)
-            if (!jsonOutput)
-            {
-                AnsiConsole.MarkupLine("[cyan]🎤 Recording... Press Enter to stop.[/]");
-            }
-
-            // Execute command with configured timeout
-            var command = new TenSecondTom.Features.Audio.Commands.RecordCommand
-            {
-                AudioConfig = ConvertSttSelectionToConfig(sttSelection, audioConfig),
-                MaxDurationSeconds = audioConfig.Timeouts.RecordSeconds  // Use configured timeout
-            };
-
-            var result = await handler.Handle(command, CancellationToken.None);
-
-            if (result.IsSuccess && result.Value is not null)
-            {
-                var recording = result.Value;
-
-                if (jsonOutput)
-                {
-                    var output = new
-                    {
-                        success = true,
-                        audio_path = recording.AudioFilePath,
-                        transcription_path = recording.TranscriptionFilePath,
-                        text = File.ReadAllText(recording.TranscriptionFilePath),
-                        duration_seconds = recording.Duration.TotalSeconds,
-                        word_count = recording.TranscriptionWordCount,
-                        stt_engine = recording.SttEngine.ToString(),
-                        stt_model = recording.SttModel,
-                        recorded_at = recording.RecordedAt,
-                        file_size_bytes = recording.FileSizeBytes
-                    };
-                    Console.WriteLine(System.Text.Json.JsonSerializer.Serialize(output, SnakeCaseJsonOptions));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[green]✓[/] Recording complete ({recording.Duration.TotalSeconds:F1}s)");
-                    AnsiConsole.MarkupLine($"[green]✓[/] Transcription complete ({recording.TranscriptionWordCount} words, {recording.SttEngine})");
-                    AnsiConsole.WriteLine();
-                    AnsiConsole.MarkupLine("[bold]Transcript:[/]");
-
-                    // Read and display the transcription text (strip YAML frontmatter)
-                    var transcriptContent = File.ReadAllText(recording.TranscriptionFilePath);
-                    var lines = transcriptContent.Split('\n');
-                    bool inFrontmatter = false;
-                    var transcriptText = new System.Text.StringBuilder();
-                    foreach (var line in lines)
-                    {
-                        if (line.Trim() == "---")
-                        {
-                            inFrontmatter = !inFrontmatter;
-                            continue;
-                        }
-                        if (!inFrontmatter && !string.IsNullOrWhiteSpace(line))
-                        {
-                            transcriptText.AppendLine(line);
-                        }
-                    }
-
-                    AnsiConsole.MarkupLine($"[dim]{transcriptText.ToString().Trim().EscapeMarkup()}[/]");
-                    AnsiConsole.WriteLine();
-                    AnsiConsole.MarkupLine($"[dim]Audio saved: {recording.AudioFilePath.EscapeMarkup()}[/]");
-                }
-                return 0;
-            }
-            else
-            {
-                if (jsonOutput)
-                {
-                    Console.WriteLine(System.Text.Json.JsonSerializer.Serialize(new
-                    {
-                        success = false,
-                        error = result.Error ?? "Recording failed"
-                    }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] Recording failed: {(result.Error ?? "Unknown error").EscapeMarkup()}");
-                }
-                return 1;
-            }
-        });
-
-        return recordCommand;
+        return RecordCommandBuilder.BuildRecordCommand(serviceProvider, jsonOutputOption);
     }
 
     private static Command BuildGenerateCommand(IServiceProvider serviceProvider, Option<bool> jsonOutputOption)
@@ -752,26 +551,12 @@ public static class CommandRegistry
 
             if (result.IsSuccess)
             {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = true, message = "Setup completed successfully" }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine("[green]✓[/] Setup completed successfully!");
-                }
+                CommandOutputFormatter.WriteSuccess("Setup completed successfully!", jsonOutput);
                 return 0;
             }
             else
             {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = false, error = result.Error }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] Setup failed: {result.Error.EscapeMarkup()}");
-                }
+                CommandOutputFormatter.WriteError($"Setup failed: {result.Error}", jsonOutput);
                 return 1;
             }
         });
@@ -781,405 +566,7 @@ public static class CommandRegistry
 
     private static Command BuildConfigCommand(IServiceProvider serviceProvider, Option<bool> jsonOutputOption)
     {
-        var configCommand = new Command("config", "View and manage Ten Second Tom configuration");
-
-        // Show subcommand
-        var showCommand = new Command("show", "Display current configuration");
-        var showSecretsOption = new Option<bool>("--show-secrets")
-        {
-            Description = "Show full API keys (last 4 characters by default)"
-        };
-        showCommand.Options.Add(showSecretsOption);
-        showCommand.Options.Add(jsonOutputOption);
-
-        showCommand.SetAction(async (parseResult) =>
-        {
-            bool showSecrets = parseResult.GetValue(showSecretsOption);
-            bool jsonOutput = parseResult.GetValue(jsonOutputOption);
-
-            var handler = serviceProvider.GetRequiredService<ConfigCommandHandler>();
-            
-            var command = new ConfigCommand
-            {
-                Action = ConfigAction.Show,
-                SettingName = null,
-                SettingValue = null,
-                ShowSecrets = showSecrets
-            };
-
-            var result = await handler.Handle(command, CancellationToken.None).ConfigureAwait(false);
-
-            if (result.IsSuccess)
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(result.Value));
-                }
-                else
-                {
-                    DisplayConfiguration(result.Value!, showSecrets);
-                }
-                return 0;
-            }
-            else
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = false, error = result.Error }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] {result.Error.EscapeMarkup()}");
-                }
-                return 1;
-            }
-        });
-
-        // LLM subcommand - interactive configuration for LLM provider and model
-        var setCommand = new Command("set", "Update a configuration setting");
-        var settingNameArg = new Argument<string>("setting")
-        {
-            Description = "Setting name (llm-provider, api-key, memory-directory, ssh-key-path, log-level, retention-days)"
-        };
-        var settingValueArg = new Argument<string>("value")
-        {
-            Description = "New value for the setting"
-        };
-        
-        setCommand.Arguments.Add(settingNameArg);
-        setCommand.Arguments.Add(settingValueArg);
-        setCommand.Options.Add(jsonOutputOption);
-
-        setCommand.SetAction(async (parseResult) =>
-        {
-            string settingName = parseResult.GetValue(settingNameArg)!;
-            string settingValue = parseResult.GetValue(settingValueArg)!;
-            bool jsonOutput = parseResult.GetValue(jsonOutputOption);
-
-            var handler = serviceProvider.GetRequiredService<ConfigCommandHandler>();
-            
-            var command = new ConfigCommand
-            {
-                Action = ConfigAction.Set,
-                SettingName = settingName,
-                SettingValue = settingValue,
-                ShowSecrets = false
-            };
-
-            var result = await handler.Handle(command, CancellationToken.None).ConfigureAwait(false);
-
-            if (result.IsSuccess)
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = true, message = $"Updated {settingName}" }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[green]✓[/] Updated [yellow]{settingName.EscapeMarkup()}[/] successfully");
-                }
-                return 0;
-            }
-            else
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = false, error = result.Error }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] {result.Error.EscapeMarkup()}");
-                }
-                return 1;
-            }
-        });
-
-        // LLM subcommand - interactive configuration for LLM provider and model
-        var llmCommand = new Command("llm", "Configure LLM provider and model interactively");
-        llmCommand.Options.Add(jsonOutputOption);
-
-        llmCommand.SetAction(async (parseResult) =>
-        {
-            bool jsonOutput = parseResult.GetValue(jsonOutputOption);
-
-            var handler = serviceProvider.GetRequiredService<ConfigCommandHandler>();
-            
-            var command = new ConfigCommand
-            {
-                Action = ConfigAction.Set,
-                SettingName = "llm",
-                SettingValue = null,
-                ShowSecrets = false
-            };
-
-            var result = await handler.Handle(command, CancellationToken.None).ConfigureAwait(false);
-
-            if (result.IsSuccess)
-            {
-                if (jsonOutput)
-                {
-                    var config = result.Value!;
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new 
-                    { 
-                        success = true, 
-                        provider = config.Llm.Provider.ToString(),
-                        model = config.Llm.Model
-                    }));
-                }
-                // Success message already displayed by handler
-                return 0;
-            }
-            else
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = false, error = result.Error }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] {result.Error.EscapeMarkup()}");
-                }
-                return 1;
-            }
-        });
-
-        // Reset subcommand
-        var validateCommand = new Command("validate", "Validate current configuration");
-        validateCommand.Options.Add(jsonOutputOption);
-
-        validateCommand.SetAction(async (parseResult) =>
-        {
-            bool jsonOutput = parseResult.GetValue(jsonOutputOption);
-
-            var handler = serviceProvider.GetRequiredService<ConfigCommandHandler>();
-            
-            var command = new ConfigCommand
-            {
-                Action = ConfigAction.Validate,
-                SettingName = null,
-                SettingValue = null,
-                ShowSecrets = false
-            };
-
-            var result = await handler.Handle(command, CancellationToken.None).ConfigureAwait(false);
-
-            if (result.IsSuccess)
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = true, message = "Configuration is valid" }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine("[green]✓[/] Configuration is valid");
-                }
-                return 0;
-            }
-            else
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = false, error = result.Error }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] {result.Error.EscapeMarkup()}");
-                }
-                return 1;
-            }
-        });
-
-        // Audio subcommand - interactive configuration for audio recording and processing
-        var audioCommand = new Command("audio", "Configure audio recording and processing settings interactively");
-        audioCommand.Options.Add(jsonOutputOption);
-
-        audioCommand.SetAction(async (parseResult) =>
-        {
-            bool jsonOutput = parseResult.GetValue(jsonOutputOption);
-
-            var handler = serviceProvider.GetRequiredService<ConfigCommandHandler>();
-
-            var command = new ConfigCommand
-            {
-                Action = ConfigAction.Set,
-                SettingName = "audio",
-                SettingValue = null,
-                ShowSecrets = false
-            };
-
-            var result = await handler.Handle(command, CancellationToken.None).ConfigureAwait(false);
-
-            if (result.IsSuccess)
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new
-                    {
-                        success = true,
-                        message = "Audio configuration updated successfully"
-                    }));
-                }
-                // Success message already displayed by handler
-                return 0;
-            }
-            else
-            {
-                if (jsonOutput)
-                {
-                    AnsiConsole.WriteLine(System.Text.Json.JsonSerializer.Serialize(new { success = false, error = result.Error }));
-                }
-                else
-                {
-                    AnsiConsole.MarkupLine($"[red]✗[/] {result.Error.EscapeMarkup()}");
-                }
-                return 1;
-            }
-        });
-
-        configCommand.Subcommands.Add(showCommand);
-        configCommand.Subcommands.Add(setCommand);
-        configCommand.Subcommands.Add(llmCommand);
-        configCommand.Subcommands.Add(audioCommand);
-        configCommand.Subcommands.Add(validateCommand);
-
-        return configCommand;
-    }
-
-    private static void DisplayConfiguration(ConfigurationSettings config, bool showSecrets)
-    {
-        var table = new Spectre.Console.Table()
-            .Border(TableBorder.Rounded)
-            .AddColumn("[yellow]Setting[/]")
-            .AddColumn("[yellow]Value[/]");
-
-        // SSH Configuration
-        table.AddRow("SSH Key Path", config.Ssh.KeyPath?.EscapeMarkup() ?? "[dim]Not set[/]");
-        table.AddRow("SSH Key Source", config.Ssh.KeySource?.ToString() ?? "[dim]Not set[/]");
-
-        // LLM Configuration
-        table.AddRow("LLM Provider", config.Llm.Provider.ToString());
-        
-        // Model - show friendly name if available
-        string modelDisplay = "[dim]Not set[/]";
-        if (!string.IsNullOrEmpty(config.Llm.Model))
-        {
-            var model = ModelRegistry.GetById(config.Llm.Model);
-            modelDisplay = model != null 
-                ? $"{model.DisplayName.EscapeMarkup()} ({model.CostTier})"
-                : config.Llm.Model.EscapeMarkup();
-        }
-        table.AddRow("Model", modelDisplay);
-        
-        string apiKeyDisplay = showSecrets 
-            ? config.Llm.ApiKey ?? "[dim]Not set[/]"
-            : MaskApiKey(config.Llm.ApiKey);
-        table.AddRow("API Key", apiKeyDisplay);
-
-        // Storage Configuration
-        table.AddRow("Memory Directory", config.Storage.MemoryDirectory?.EscapeMarkup() ?? "[dim]Not set[/]");
-
-        // Optional Configuration
-        table.AddRow("Log Level", config.Optional.LogLevel.ToString());
-        table.AddRow("Retention Days", config.Optional.RetentionDays.ToString(System.Globalization.CultureInfo.InvariantCulture));
-
-        // Audio Configuration
-        table.AddRow("[yellow]Audio Settings[/]", "");
-        table.AddRow("  STT Provider", config.Audio.SttProvider);
-        table.AddRow("  Keep Files", config.Audio.KeepFiles ? "Yes" : "No");
-        
-        // Audio Recorder Configuration
-        table.AddRow("  [dim]Recorder:[/]", "");
-        table.AddRow("    Input Volume", config.Audio.Recorder.InputVolume.ToString("0.0", System.Globalization.CultureInfo.InvariantCulture));
-        table.AddRow("    Noise Reduction", config.Audio.Recorder.EnableNoiseReduction ? "Enabled" : "Disabled");
-        table.AddRow("    Frequency Filters", config.Audio.Recorder.EnableFrequencyFilters ? "Enabled" : "Disabled");
-        
-        // Audio Preprocessing Configuration
-        table.AddRow("  [dim]Preprocessing:[/]", "");
-        table.AddRow("    Remove Silence", config.Audio.Preprocessing.RemoveSilence ? "Enabled" : "Disabled");
-        table.AddRow("    Silence Threshold", $"{config.Audio.Preprocessing.SilenceThresholdDb} dB");
-        table.AddRow("    Min Silence Duration", $"{config.Audio.Preprocessing.MinimumSilenceDurationMs} ms");
-
-        // Metadata
-        table.AddRow("[dim]Created[/]", $"[dim]{config.CreatedAt:yyyy-MM-dd HH:mm:ss}[/]");
-        if (config.LastModifiedAt.HasValue)
-        {
-            table.AddRow("[dim]Modified[/]", $"[dim]{config.LastModifiedAt.Value:yyyy-MM-dd HH:mm:ss}[/]");
-        }
-
-        AnsiConsole.Write(table);
-    }
-
-    private static string MaskApiKey(string? apiKey)
-    {
-        if (string.IsNullOrWhiteSpace(apiKey))
-            return "[dim]Not set[/]";
-
-        if (apiKey.Length <= 4)
-            return "••••";
-
-        return $"••••{apiKey[^4..]}";
-    }
-
-    /// <summary>
-    /// Converts CLI SttSelection intent to AudioConfiguration.
-    /// Maps user-friendly CLI options (auto/local/openai) to the proper configuration.
-    /// </summary>
-    /// <param name="selection">The STT selection from CLI (auto, local, or openai)</param>
-    /// <param name="baseConfig">The base audio configuration from appsettings</param>
-    /// <returns>AudioConfiguration with appropriate provider and fallback settings</returns>
-    private static Configuration.AudioConfiguration ConvertSttSelectionToConfig(TenSecondTom.Features.Audio.Models.SttSelection selection, Configuration.AudioConfiguration baseConfig)
-    {
-        return selection switch
-        {
-            // Auto: Try local provider first, fallback to OpenAI cloud if enabled
-            TenSecondTom.Features.Audio.Models.SttSelection.Auto => new Configuration.AudioConfiguration
-            {
-                SttProvider = baseConfig.SttProvider,
-                SttBinaryPath = baseConfig.SttBinaryPath,
-                SttModel = baseConfig.SttModel,
-                SttApiKey = baseConfig.SttApiKey,
-                SttFallbackEnabled = true,
-                SttFallbackProvider = baseConfig.SttFallbackProvider,
-                SttFallbackBinaryPath = baseConfig.SttFallbackBinaryPath,
-                SttFallbackModel = baseConfig.SttFallbackModel,
-                SttFallbackApiKey = baseConfig.SttFallbackApiKey,
-                Recorder = baseConfig.Recorder,
-                KeepFiles = baseConfig.KeepFiles,
-                Preprocessing = baseConfig.Preprocessing,
-                Timeouts = baseConfig.Timeouts
-            },
-
-            // Local: Use only the configured local provider (whisper.cpp, ollama, etc.) - no fallback
-            TenSecondTom.Features.Audio.Models.SttSelection.Local => new Configuration.AudioConfiguration
-            {
-                SttProvider = baseConfig.SttProvider,
-                SttBinaryPath = baseConfig.SttBinaryPath,
-                SttModel = baseConfig.SttModel,
-                SttApiKey = baseConfig.SttApiKey,
-                SttFallbackEnabled = false,
-                Recorder = baseConfig.Recorder,
-                KeepFiles = baseConfig.KeepFiles,
-                Preprocessing = baseConfig.Preprocessing,
-                Timeouts = baseConfig.Timeouts
-            },
-
-            // OpenAI: Force OpenAI provider, no fallback
-            TenSecondTom.Features.Audio.Models.SttSelection.OpenAI => new Configuration.AudioConfiguration
-            {
-                SttProvider = SttProviders.OpenAI,
-                SttBinaryPath = baseConfig.SttBinaryPath,
-                SttModel = baseConfig.SttModel,
-                SttApiKey = baseConfig.SttApiKey,
-                SttFallbackEnabled = false,
-                Recorder = baseConfig.Recorder,
-                KeepFiles = baseConfig.KeepFiles,
-                Preprocessing = baseConfig.Preprocessing,
-                Timeouts = baseConfig.Timeouts
-            },
-
-            _ => throw new ArgumentOutOfRangeException(nameof(selection), selection,
-                $"Unsupported STT selection: {selection}")
-        };
+        return ConfigCommandBuilder.BuildConfigCommand(serviceProvider, jsonOutputOption);
     }
 }
